progress on giztimer

+++ src/lostandfound/qrand.cpp
@@ -1,8 +1,8 @@
 #include "qrand.h"
 
-int qseed = 0x3039;
+i32 qseed = 0x3039;
 
-u16 qrand(void) {
+i32 qrand(void) {
     qseed = qseed * 0x24cd + 1 & 0xffff;
 
     return qseed;
